requirements.txt

streamlit==1.26.0
deeplake==3.6.19
openai==0.27.6
langchain==0.0.267
tiktoken==0.4.0
unstructured==0.6.5
pdf2image==1.16.3
pytesseract==0.3.10
beautifulsoup4==4.12.2
bs4==0.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
tqdm==4.65.0
GitPython==3.1.31
pypdf==3.15.1
